Fix all warnings in the main codebase flagged by -Wsigned-compare
Remember, the code
   int is_less_than(int a, unsigned b) {
      return a < b;
   }
is buggy, since the C integer promotion rules basically turn it into
   int is_less_than(int a, unsigned b) {
      return ((unsigned)a) < b;
   }
and we really want something closer to
   int is_less_than(int a, unsigned b) {
      return a < 0 || ((unsigned)a) < b;
   }
.

Suggested by an example from Ralph Castain

+++ select.c
@@ -234,11 +234,14 @@ select_add(struct event_base *base, int fd, short old, short events, void *p)
 	if (sop->event_fds < fd) {
 		int fdsz = sop->event_fdsz;
 
-		if (fdsz < sizeof(fd_mask))
-			fdsz = sizeof(fd_mask);
+		if (fdsz < (int)sizeof(fd_mask))
+			fdsz = (int)sizeof(fd_mask);
 
+		/* In theory we should worry about overflow here.  In
+		 * reality, though, the highest fd on a unixy system will
+		 * not overflow here. XXXX */
 		while (fdsz <
-		    (howmany(fd + 1, NFDBITS) * sizeof(fd_mask)))
+		    (int) (howmany(fd + 1, NFDBITS) * sizeof(fd_mask)))
 			fdsz *= 2;
 
 		if (fdsz != sop->event_fdsz) {
